Fix a few issues
is_subclass_of() must be called with Throwable, not Exception, or passing an Exception won't work.

Also, we cannot check for our own static methods in __callStatic(), as those will be called directly.

Instead of using a static property, it's just easier to avoid having an state in a purely static class, and therefore catching any exception (shall it occur), and "casting" it to whatever exception we need.